+++ include/boost/beast/core/detail/bind_handler.hpp
@@ -7,8 +7,8 @@
 // Official repository: https://github.com/boostorg/beast
 //
 
-#ifndef BOOST_BEAST_BIND_DETAIL_HANDLER_HPP
-#define BOOST_BEAST_BIND_DETAIL_HANDLER_HPP
+#ifndef BOOST_BEAST_DETAIL_BIND_HANDLER_HPP
+#define BOOST_BEAST_DETAIL_BIND_HANDLER_HPP
 
 #include <boost/beast/core/detail/integer_sequence.hpp>
 #include <boost/asio/handler_alloc_hook.hpp>