+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Version 86:
 
 * Boost prep
 * Remove use of lexical_cast
+* Use custom variant
 
 --------------------------------------------------------------------------------